Moved account service related code into separate folder
