+++ api/pacman-api-admin/src/main/java/com/tmobile/pacman/api/admin/controller/AccountsController.java
@@ -10,8 +10,7 @@ import com.tmobile.pacman.api.admin.domain.Response;
 import com.tmobile.pacman.api.admin.exceptions.PluginServiceException;
 import com.tmobile.pacman.api.admin.factory.AccountFactory;
 import com.tmobile.pacman.api.admin.factory.PluginFactory;
-import com.tmobile.pacman.api.admin.repository.model.AccountDetails;
-import com.tmobile.pacman.api.admin.repository.service.AccountsService;
+import com.tmobile.pacman.api.admin.repository.service.accounts.AccountsService;
 import com.tmobile.pacman.api.admin.repository.service.AssetGroupService;
 import com.tmobile.pacman.api.admin.repository.service.PluginsService;
 import com.tmobile.pacman.api.admin.repository.service.UserPreferencesService;
@@ -24,7 +23,6 @@ import io.swagger.annotations.ApiParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -41,7 +39,6 @@ import org.springframework.web.bind.annotation.RestController;
 import java.security.Principal;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static com.tmobile.pacman.api.admin.common.AdminConstants.UNEXPECTED_ERROR_OCCURRED;
