+++ api/pacman-api-admin/src/main/java/com/tmobile/pacman/api/admin/repository/service/accounts/ContrastAccountServiceImpl.java
@@ -13,7 +13,7 @@
  *  License for the specific language governing permissions and limitations under
  *  the License.
  ******************************************************************************/
-package com.tmobile.pacman.api.admin.repository.service;
+package com.tmobile.pacman.api.admin.repository.service.accounts;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicSessionCredentials;
