+++ api/pacman-api-admin/src/main/java/com/tmobile/pacman/api/admin/repository/service/accounts/GcpAccountServiceImpl.java
@@ -1,4 +1,4 @@
-package com.tmobile.pacman.api.admin.repository.service;
+package com.tmobile.pacman.api.admin.repository.service.accounts;
 
 import com.amazonaws.SdkClientException;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
@@ -37,13 +37,12 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.io.*;
-import java.nio.file.Files;
 import java.security.GeneralSecurityException;
 import java.util.List;
 import java.util.Map;
 
 @Service
-public class GcpAccountServiceImpl extends AbstractAccountServiceImpl implements AccountsService{
+public class GcpAccountServiceImpl extends AbstractAccountServiceImpl implements AccountsService {
 
     private static final Logger LOGGER= LoggerFactory.getLogger(GcpAccountServiceImpl.class);
     public static final String MISSING_MANDATORY_PARAMETER = "Missing mandatory parameter: ";
